% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{begin_processing}
\alias{begin_processing}
\title{Runs the pre-processing step on a startr project.}
\usage{
begin_processing(should_clean_processing_variables = TRUE)
}
\arguments{
\item{should_clean_processing_variables}{Either TRUE, FALSE, or pulled from the environment if set.}
}
\value{
A list of all environment variables present before the function was run
}
\description{
The pre-processing step, run as part of \code{upstartr::\link[upstartr:run_process]{run_process}}
during the \code{process.R} stage of a startr project, logs all variables
currently in the global environment, which will then be removed during the
post-processing step to keep the startr environment unpolluted.
}
