% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_index}
\alias{calc_index}
\title{Index values}
\usage{
calc_index(m, base = 0)
}
\arguments{
\item{m}{Numeric vector to index to first value.}

\item{base}{Base to index against. (Default: 0)}
}
\value{
An vector of indexed values.
}
\description{
Index numeric vector to first value. By default, the index base will be 0,
turning regular values into percentage change. In some cases, you may want to index
to a different base, like 100, such as if you're looking at financial data.
}
\examples{
calc_index(c(5, 2, 8, 17, 7, 3, 1, -4))
calc_index(c(5, 2, 8, 17, 7, 3, 1, -4), base = 100)

}
