% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_columns}
\alias{clean_columns}
\title{Cleans up column names by forcing them into tidyverse style}
\usage{
clean_columns(x)
}
\arguments{
\item{x}{A vector of column names.}
}
\value{
A character vector of column names.
}
\description{
Zero-configuration function that takes unwieldy column names and
coerces them into tidyverse-styled column names.
}
\examples{
clean_columns(c("Date of Purchase", "Item No.", "description", "",
  "Transaction at Jane's Counter?", "Auditing - Worth it?"))

}
