% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{combine_csvs}
\alias{combine_csvs}
\title{Combine CSVs in a directory}
\usage{
combine_csvs(dir, pattern = "*.csv", ...)
}
\arguments{
\item{dir}{Path to the directory to look at for files.}

\item{pattern}{Pattern to use for detecting files. (Default: '*.csv')}

\item{...}{Parameters to pass to \code{readr::\link[readr:read_csv]{read_csv}}.}
}
\value{
A tibble of concatenated data from multiple CSV files.
}
\description{
Given a directory (and, optionally, a pattern to search against),
concatenate all CSV files into a single tibble.
}
