% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{combine_excels}
\alias{combine_excels}
\title{Combine Excel files in a directory}
\usage{
combine_excels(dir, pattern = "*.xls[x]?", all_sheets = FALSE, ...)
}
\arguments{
\item{dir}{Path to the directory to look at for files.}

\item{pattern}{Pattern to use for detecting files. (Default: '*.xls[x]?')}

\item{all_sheets}{Should this function also concatenate all sheets within each
Excel file into one long tibble? (Default: FALSE)}

\item{...}{Parameters to pass to \code{readxl::\link[readxl:read_excel]{read_excel}}.}
}
\value{
A tibble of concatenated data from multiple Excel files.
}
\description{
Given a directory (and, optionally, a pattern to search against),
concatenate all Excel files into a single tibble.
}
