% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_data_out}
\alias{dir_data_out}
\title{Get path within disposable data outputs directory.}
\usage{
dir_data_out(...)
}
\arguments{
\item{...}{Any number of path strings, passed in the same fashion as \code{here::\link[here:here]{here}}.}
}
\value{
A path string.
}
\description{
Constructs a path within startr's data/out/ directory.
}
