% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_data_raw}
\alias{dir_data_raw}
\title{Get path within raw data directory.}
\usage{
dir_data_raw(...)
}
\arguments{
\item{...}{Any number of path strings, passed in the same fashion as \code{here::\link[here:here]{here}}.}
}
\value{
A path string.
}
\description{
Constructs a path within startr's data/raw/ directory.
}
