% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_path}
\alias{dir_path}
\title{Construct an arbitrary path.}
\usage{
dir_path(...)
}
\arguments{
\item{...}{Any number of path strings, passed in the same fashion as \code{here::\link[here:here]{here}}.}
}
\value{
A path string.
}
\description{
Convenience function that constructs a path. Wraps \code{here::\link[here:here]{here}}.
}
