% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_plots}
\alias{dir_plots}
\title{Get path within plots directory.}
\usage{
dir_plots(...)
}
\arguments{
\item{...}{Any number of path strings, passed in the same fashion as \code{here::\link[here:here]{here}}.}
}
\value{
A path string.
}
\description{
Constructs a path within startr's plots/ directory.
}
