% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_reports}
\alias{dir_reports}
\title{Get path within reports directory.}
\usage{
dir_reports(...)
}
\arguments{
\item{...}{Any number of path strings, passed in the same fashion as \code{here::\link[here:here]{here}}.}
}
\value{
A path string.
}
\description{
Constructs a path within startr's reports/ directory.
}
