% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_scrape}
\alias{dir_scrape}
\title{Get path within scrape directory.}
\usage{
dir_scrape(...)
}
\arguments{
\item{...}{Any number of path strings, passed in the same fashion as \code{here::\link[here:here]{here}}.}
}
\value{
A path string.
}
\description{
Constructs a path within startr's scrape/ directory.
}
