% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_src}
\alias{dir_src}
\title{Get path within src directory}
\usage{
dir_src(...)
}
\arguments{
\item{...}{Any number of path strings, passed in the same fashion as \code{here::\link[here:here]{here}}.}
}
\value{
A path string.
}
\description{
Constructs a path within startr's main R/ directory.
}
