% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{end_processing}
\alias{end_processing}
\title{Runs the post-processing step on a startr project.}
\usage{
end_processing(
  should_clean_processing_variables = TRUE,
  should_beep = TRUE,
  logged_vars = NULL
)
}
\arguments{
\item{should_clean_processing_variables}{Either TRUE, FALSE, or pulled from the environment if set.}

\item{should_beep}{Either TRUE, FALSE, or pulled from the environment if set.}

\item{logged_vars}{A list of variables that existed before the processing step began.}
}
\value{
No return value, called for side effects
}
\description{
The post-processing step, run as part of \code{upstartr::\link[upstartr:run_process]{run_process}}
during the \code{process.R} stage of a startr project, removes all variables
saved by \code{upstartr::\link[upstartr:begin_processing]{begin_processing}}
and then beeps to announce it's finished.
}
