% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Opposite of \%in\%}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{The vector you want to check.}

\item{table}{Table in which to do lookups against x.}
}
\value{
Same form of return as \%in\% — except it will return only elements on the lhs that aren't present on the rhs
}
\description{
Given vectors A and B, returns only the entities from vector A that don't occur in vector B.
}
\examples{
c(1, 2, 3, 4, 5) \%not_in\% c(4, 5, 6, 7, 8)

}
