% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index}
\alias{index}
\title{Index values}
\usage{
index(m, base = 0)
}
\arguments{
\item{m}{Numeric vector to index to first value.}

\item{base}{Base to index against. (Default: 0)}
}
\value{
The mode of that vector.
}
\description{
Index numeric vector to first value. By default, the index base will be 0,
turning regular values into percentage change. In some cases, you may want to index
to a different base, like 100, such as if you're looking at financial data.
}
