% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode}
\alias{mode}
\title{Calculate mode}
\usage{
mode(x)
}
\arguments{
\item{x}{Any kind of vector — numeric, character, logical.}
}
\value{
The mode of that vector.
}
\description{
Calculates the mode of a given vector.
}
