% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not.na}
\alias{not.na}
\title{Opposite of is.na}
\usage{
not.na(x)
}
\arguments{
\item{x}{A vector to check for NAs against.}
}
\description{
Given a vector, returns TRUE for all entities that aren't NA.
}
\examples{
not.na(c(1, NA, 2, NA))

}
