% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_all_excel_sheets}
\alias{read_all_excel_sheets}
\title{Combine all sheets in an Excel file}
\usage{
read_all_excel_sheets(filepath, ...)
}
\arguments{
\item{filepath}{Path to the Excel file.}

\item{...}{Parameters to pass to \code{readxl::\link[readxl:read_excel]{read_excel}}.}
}
\value{
A tibble data concatenated from a all sheets in an Excel file.
}
\description{
Reads all sheets in a single Excel file using \code{readxl::\link[readxl:read_excel]{read_excel}}
and concatenates them into a single, long tibble.
}
