% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{render_notebook}
\alias{render_notebook}
\title{Renders out an RMarkdown notebook.}
\usage{
render_notebook(notebook_file, output_dir = dir_reports())
}
\arguments{
\item{notebook_file}{The path for the RMarkdown notebook you're rendering.}

\item{output_dir}{The directory to write the outputs to.}
}
\value{
No return value, called for side effects
}
\description{
Renders an RMarkdown notebook using \code{upstartr::\link[upstartr:render_notebook]{render_notebook}}
and then beeps.
}
