% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_config}
\alias{run_config}
\title{Configures an existing startr project}
\usage{
run_config()
}
\value{
No return value, called for side effects
}
\description{
Sources \code{config.R} and \code{functions.R} in turn.
}
