% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_notebook}
\alias{run_notebook}
\title{Runs the notebook rendering step for a startr project.}
\usage{
run_notebook(
  filename = "notebook.Rmd",
  should_beep = TRUE,
  should_render_notebook = TRUE
)
}
\arguments{
\item{filename}{The filename for the RMarkdown notebook you want to render.}

\item{should_beep}{Either TRUE, FALSE, or pulled from the environment if set.}

\item{should_render_notebook}{Either TRUE, FALSE, or pulled from the environment if set.}
}
\value{
No return value, called for side effects
}
\description{
Renders an RMarkdown notebook using \code{upstartr::\link[upstartr:render_notebook]{render_notebook}}
and then beeps.
}
