% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_process}
\alias{run_process}
\title{Runs the processing step on a startr project.}
\usage{
run_process(should_process_data = TRUE)
}
\arguments{
\item{should_process_data}{Either TRUE, FALSE, or pulled from the environment if set.}
}
\value{
No return value, called for side effects
}
\description{
Runs the pre-processing step (see \code{upstartr::\link[upstartr:begin_processing]{begin_processing}}
for details), then sources \code{process.R}, then runs the post-processing step
(see \code{upstartr::\link[upstartr:end_processing]{end_processing}} for details).
}
