% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_visualize}
\alias{run_visualize}
\title{Runs the visualization step for a startr project.}
\usage{
run_visualize()
}
\value{
No return value, called for side effects
}
\description{
Sources \code{visualize.R}.
}
