% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scale_x_percent}
\alias{scale_x_percent}
\title{Create a continuous x-axis scale using percentages}
\usage{
scale_x_percent(...)
}
\arguments{
\item{...}{All your usual continuous x-axis scale parameters.}
}
\value{
A scale object to be consumed by ggplot2.
}
\description{
Convenience function to return a scale_x_continuous function using percentage labels.
}
