% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scale_y_percent}
\alias{scale_y_percent}
\title{Create a continuous y-axis scale using percentages}
\usage{
scale_y_percent(...)
}
\arguments{
\item{...}{All your usual continuous y-axis scale parameters.}
}
\value{
A scale object to be consumed by ggplot2.
}
\description{
Convenience function to return a scale_y_continuous function using percentage labels.
}
