% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scale_y_reordered}
\alias{scale_y_reordered}
\title{Passed after reorder_within to reorder x-axis along facets}
\description{
See \code{tidytext::\link[tidytext:scale_y_reordered]{scale_y_reordered}} for details.
}
\keyword{internal}
