% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unaccent}
\alias{unaccent}
\title{De-accents strings}
\usage{
unaccent(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of strings without accents.
}
\description{
Replace accented characters with their non-accented versions. Useful when dealing
with languages like French, Spanish or Portuguese, where accents can lead to
compatibility issues during data analysis.
}
