% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_plot}
\alias{write_plot}
\title{Write out a ggplot2 graphic with minimal configuration}
\usage{
write_plot(variable, output_dir = dir_plots(), ...)
}
\arguments{
\item{variable}{A tibble or dataframe object.}

\item{output_dir}{The directory to save the plot out to.}

\item{...}{Other settings to pass to ggsave, such as format, width, height or dpi.}
}
\value{
No return value, called for side effects
}
\description{
Takes a \code{ggplot2} object and writes it to disk via \code{ggplot2::\link[ggplot2:ggsave]{ggsave}} using the
variable name as the filename.
}
