% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_shp}
\alias{write_shp}
\title{Write a shapefile to disk}
\usage{
write_shp(shp, path, ...)
}
\arguments{
\item{shp}{A spatial object.}

\item{path}{The desired filepath for the shapefile.}

\item{...}{Other settings to pass to st_write, such as format, width, height or dpi.}
}
\value{
No return value, called for side effects
}
\description{
Utility function that wraps \code{sf::\link[sf:st_write]{st_write}}, but first
removes a previous version of the shapefile if it exists (by default, \code{sf::\link[sf:st_write]{st_write}}
will throw an error.)
}
