% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate mode}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{Any kind of vector — numeric, character, logical.}
}
\value{
The mode(s) of that vector.
}
\description{
Calculates the mode of a given vector.
}
\examples{
calc_mode(c(1, 1, 2, 3, 4))
calc_mode(c('the', 'quick', 'brown', 'fox', 'jumped', 'over', 'the', 'lazy', 'dog'))
calc_mode(c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))

}
