% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{not.null}
\alias{not.null}
\title{Opposite of is.null}
\usage{
not.null(x)
}
\arguments{
\item{x}{A vector to check for NULLs against.}
}
\value{
Elements that aren't NULL
}
\description{
Given a list, returns TRUE for all entities that aren't NULL.
}
\examples{
not.null(list(1, NULL, 2, NULL))

}
