% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_non_utf8}
\alias{remove_non_utf8}
\title{Removes non-UTF-8 characters}
\usage{
remove_non_utf8(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of strings without non-UTF-8 characters.
}
\description{
Removes non-UTF-8 characters in a given character vector.
}
\examples{
non_utf8 <- 'fa\xE7ile'
  Encoding(non_utf8) <- 'latin1'
  remove_non_utf8(non_utf8)

}
