% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unaccent}
\alias{unaccent}
\title{De-accents strings}
\usage{
unaccent(x, remove.nonconverted = FALSE, ...)
}
\arguments{
\item{x}{A character vector.}

\item{remove.nonconverted}{Should the function remove unmapped encodings? (Default: FALSE)}

\item{...}{Parameters passed to \code{textclean::\link[textclean:replace_non_ascii]{replace_non_ascii}}}
}
\value{
A character vector of strings without accents.
}
\description{
Replace accented characters with their non-accented versions. Useful when dealing
with languages like French, Spanish or Portuguese, where accents can lead to
compatibility issues during data analysis.
}
\examples{
unaccent('façile')
unaccent('Montréal')

}
